refactor(ui): move kernel utils to shared utils directory
